+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - event.attempt sub-resource to allow retrieval of information on failed webhook event delivery attempts
 - pdf method for retrieving PDF receipts from reversed invoice logs
 - page functions as a manual-pagination alternative to queries
+- Institution resource to allow query of institutions recognized by the Brazilian Central Bank for Pix and TED transactions
 
 ## [2.3.4] - 2021-03-04
 ### Added